backend/requirements.txt
Django>=3.2
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
